setup/terraform/web/requirements.txt
flask>=1.0.2
flask-bootstrap>=3.3.7.1
flask-login>=0.4.1
flask-migrate>=2.5.2
flask-httpauth>=3.3.0
flask-wtf>=0.14.2
python-dotenv>=0.10.3
